Parser.java: extract extractParameter
extractParameters contains deeply-nested code.

It is difficult to keep track of the logic of deeply-nested code.

Let's pull up the extraction of each parameter into a new method named
parseParameter.

The for loop body in extractParameters will only contain a call to
extractParameter, which reduces the number of levels of indentation.