Added Sorting functionality to Dude.
Currently, tasks are stored such that they are ordered by when tasks are
added into dude by user. It has no order.

The ordering of the tasks are necessary as it is a relevant information
to the user.

I have modified Dude in such a way that it sorts the tasks added by
time, every time a new task is added to added it dude. The Tasks are
sorted by By-Dates (if Deadline Task) or From-Date (for event tasks).
Todo tasks are sorted by which one was last added.

Now the user will get a sorted task list everywhere in the program,
allowing the user to see tasks chronologically.